mapbox + user routers documentation, geoJsonUsersList moved to mapboxRouter
